+++ insights-ui/src/app/api/actions/tickers/[ticker]/trigger-all/route.ts
@@ -1,8 +1,9 @@
 // app/api/public-equity/all-criterion-report/route.ts
+import { getCriteria } from '@/lib/industryGroupCriteria';
 import { TickerReport } from '@/types/public-equity/ticker-report-types';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
-import { getTickerReport, getCriteria } from '@/lib/publicEquity';
+import { getTickerReport } from '@/lib/publicEquity';
 
 const PE_US_REITS_WEBHOOK_URL = process.env.PE_US_REITS_WEBHOOK_URL!;
