+++ insights-ui/src/app/api/actions/tickers/[ticker]/trigger-next/route.ts
@@ -1,9 +1,10 @@
 // app/api/public-equity/next-criterion-report/route.ts
+import { getCriteria } from '@/lib/industryGroupCriteria';
 import { ProcessingStatus } from '@/types/public-equity/ticker-report-types';
 import { NextCriterionReportRequest } from '@/types/public-equity/ticker-request-response';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
-import { getTickerReport, getCriteria } from '@/lib/publicEquity';
+import { getTickerReport } from '@/lib/publicEquity';
 
 const PE_US_REITS_WEBHOOK_URL = process.env.PE_US_REITS_WEBHOOK_URL!;
