+++ insights-ui/src/app/api/ai-criteria/route.ts
@@ -1,4 +1,10 @@
-import { generateAiCriteria, getCriteriaLookupList, getMatchingCriteriaLookupItem, updateCriteriaLookupList, uploadAiCriteriaToS3 } from '@/lib/publicEquity';
+import {
+  generateAiCriteria,
+  getCriteriaLookupList,
+  getMatchingCriteriaLookupItem,
+  updateCriteriaLookupList,
+  uploadAiCriteriaToS3,
+} from '@/lib/industryGroupCriteria';
 import { IndustryGroupCriteriaDefinition } from '@/types/public-equity/criteria-types';
 import { CreateCriteriaRequest } from '@/types/public-equity/ticker-request-response';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
